requirements.txt
beautifulsoup4==4.13.4
flake8==7.2.0
google-genai==1.16.1
lxml==5.4.0
python-dotenv==1.1.0
requests==2.32.3
